Avoid unintentional unsigned integer wraparounds

+++ src/chain.h
@@ -475,7 +475,7 @@ public:
 
     /** Return the maximal height in the chain. Is equal to chain.Tip() ? chain.Tip()->nHeight : -1. */
     int Height() const {
-        return vChain.size() - 1;
+        return (int)vChain.size() - 1;
     }
 
     /** Set/initialize a chain with a given tip. */
