+++ src/prevector.h
@@ -303,8 +303,8 @@ public:
     iterator end() { return iterator(item_ptr(size())); }
     const_iterator end() const { return const_iterator(item_ptr(size())); }
 
-    reverse_iterator rbegin() { return reverse_iterator(item_ptr(size() - 1)); }
-    const_reverse_iterator rbegin() const { return const_reverse_iterator(item_ptr(size() - 1)); }
+    reverse_iterator rbegin() { return reverse_iterator(item_ptr((int)size() - 1)); }
+    const_reverse_iterator rbegin() const { return const_reverse_iterator(item_ptr((int)size() - 1)); }
     reverse_iterator rend() { return reverse_iterator(item_ptr(-1)); }
     const_reverse_iterator rend() const { return const_reverse_iterator(item_ptr(-1)); }
